+++ Boxer/Shaders/BXShaderParametersWindowController.m
@@ -22,8 +22,6 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-@import OpenEmuShaders;
-
 #import "BXShaderParametersWindowController.h"
 #import "BXSession.h"
 #import "BXDOSWindowController.h"